feat(material): remove ThemeProvider component
This is a breaking change.  The `ThemeProvider` has been renamed to `ThemeVariantsProvider`,
and has been removed from the public API.
The Customer UI base theme is available from a `useTheme` hook,
and the Material UI theme used by components is available from a
`useMuiTheme` hook which is the Material UI `useTheme` hook re-exported under a
different name.
The provider has been renamed because the name `ThemeProvider` has come to have certain
expectations due to it's use in established component libraries and
CSS-in-JS styling libraries.